vyatta-op-vpn: prevent invalid rsa key file from being generated
If the command "generate vpn rsa-key" is aborted during key generation
it leaves behind a temporary file.  If the command is then executed
again, this temporary file is appended to rather than being replaced,
resulting in a key file with an extra : RSA { line at the beginning.

This patch checks if this temporary file exists, deleting it if it
does.

Bug #262 http://bugzilla.vyos.net/show_bug.cgi?id=262